The 2.3.1 version of the BB SDK for Smartphone
