+++ framework/src/blackberry/web/widget/Widget.java
@@ -53,18 +53,18 @@ public class Widget extends UiApplication implements GlobalEventListener {
         initialize();
         _locationURI = locationURI;
 
-		// Set our orientation
-		WidgetConfigImpl configImpl = (WidgetConfigImpl) _wConfig;
-		if (configImpl.isOrientationDefined()) {
-			int direction;
-			if (configImpl.getOrientation() == 0) {
-				direction = net.rim.device.api.system.Display.DIRECTION_PORTRAIT;   
-			} else {
-				direction = net.rim.device.api.system.Display.DIRECTION_LANDSCAPE;   
-			}
-			net.rim.device.api.ui.Ui.getUiEngineInstance().setAcceptableDirections(direction);
-		}
-		
+        // Set our orientation
+        WidgetConfigImpl configImpl = (WidgetConfigImpl) _wConfig;
+        if( configImpl.isOrientationDefined() ) {
+            int direction;
+            if( configImpl.getOrientation() == 0 ) {
+                direction = net.rim.device.api.system.Display.DIRECTION_NORTH;
+            } else {
+                direction = net.rim.device.api.system.Display.DIRECTION_LANDSCAPE;
+            }
+            net.rim.device.api.ui.Ui.getUiEngineInstance().setAcceptableDirections( direction );
+        }
+
         // Create PageManager
         PageManager pageManager = new PageManager( this, (WidgetConfigImpl) _wConfig );
 
@@ -143,9 +143,9 @@ public class Widget extends UiApplication implements GlobalEventListener {
                 waitForStartupComplete();
             }
             Widget widget = makeWidget( args, wConfig );
-            
+
             widget.enterEventDispatcher();
-			
+
             MemoryMaid mm = MemoryMaid.getInstance();
             if( mm != null ) {
                 mm.stop();
